Migrate recipe Metadata to a separate mod
