+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod cookbook;
 pub mod error;
 pub mod ingredient;
+pub mod metadata;
 pub mod recipe;
 pub mod render;
